backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-dotenv
pymongo
email-validator
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
groq
google-generativeai
scikit-learn
google-search-results
